third_party/WebKit/LayoutTests/inspector/tracing/timeline-js-samping-tracing-expected.txt

JSFrame: 211.000 / 0.226 a
JSFrame: 211.000 / 0.226 b
JSFrame: 211.000 / 0.226 c
JSFrame: 220.000 / 9.002 a
JSFrame: 220.000 / 9.002 b
JSFrame: 220.000 / 0.603 c
JSFrame: 222.000 / 3.758 d
JSFrame: 222.000 / 3.758 e
JSFrame: 228.125 / 0.877 e
JSFrame: 228.125 / 0.877 g
JSFrame: 230.000 / 0.808 f
JSFrame: 230.000 / 0.808 l
JSFrame: 230.000 / 0.808 a
